backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
pymupdf
# PDFとExcelの両方を扱えるように変更
unstructured[pdf,xlsx]
langchain-unstructured
